// lib/airstack-types.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Address: { input: any; output: any; }
  DateRange: { input: any; output: any; }
  Identity: { input: any; output: any; }
  IntString: { input: any; output: any; }
  Map: { input: any; output: any; }
  Range: { input: any; output: any; }
  Time: { input: any; output: any; }
  TimeRange: { input: any; output: any; }
};
// ...
export type Token = {
  /** Smart contract address of the token */
  address: Scalars['Address']['output'];
  /** Base URI of the token contract */
  baseURI: Maybe<Scalars['String']['output']>;
  /** Blockchain where the token smart contract is deployed */
  blockchain: Maybe<TokenBlockchain>;
  /** Unique identifier for the blockchain */
  chainId: Maybe<Scalars['String']['output']>;
  /** Token contract metadata object */
  contractMetaData: Maybe<ContractMetadata>;
  /** URI for the token's contract metadata */
  contractMetaDataURI: Maybe<Scalars['String']['output']>;
  /** The number of decimal places this token uses, default to 18 */
  decimals: Maybe<Scalars['Int']['output']>;
  /** Airstack unique identifier for the contract */
  id: Maybe<Scalars['ID']['output']>;
  /** Indicates if the token is set to be spam - true or false */
  isSpam: Maybe<Scalars['Boolean']['output']>;
  /** Block number of the token's most recent transfer */
  lastTransferBlock: Maybe<Scalars['Int']['output']>;
  /** Transaction hash of the token's most recent transfer */
  lastTransferHash: Maybe<Scalars['String']['output']>;
  /** Timestamp of the token's most recent transfer */
  lastTransferTimestamp: Maybe<Scalars['Time']['output']>;
  /** Logo image for the contract in various sizes (if available) */
  logo: Maybe<LogoSizes>;
  /** Name of the token, mirrored from the smart contract */
  name: Maybe<Scalars['String']['output']>;
  /** The owner of the token contract */
  owner: Maybe<Wallet>;
  /** off-chain data for the token project */
  projectDetails: Maybe<ProjectDetails>;
  /** Token contract metadata as it appears inside the contract */
  rawContractMetaData: Maybe<Scalars['Map']['output']>;
  /** Symbol of the token, mirrored from the smart contract */
  symbol: Maybe<Scalars['String']['output']>;
  /** Nested Query - allows querying the tokenBalance information */
  tokenBalances: Maybe<Array<TokenBalance>>;
  /** Nested Query - allows querying the tokenNFTs information */
  tokenNfts: Maybe<Array<TokenNft>>;
  /** Returns count of all NFT token attribute types and values for the given smart contract */
  tokenTraits: Maybe<Scalars['Map']['output']>;
  /** Amount of tokens in the protocol */
  totalSupply: Maybe<Scalars['String']['output']>;
  /** Token type: ERC20, ERC721, or ERC1155 */
  type: Maybe<TokenType>;
};


export type TokenTokenBalancesArgs = {
  input: InputMaybe<TokenBalancesNestedInput>;
};


export type TokenTokenNftsArgs = {
  input: InputMaybe<TokenNftsNestedInput>;
};

export type TokenBalance = {
  /** Token amount the address currently holds */
  amount: Scalars['String']['output'];
  /** Blockchain where the token smart contract is deployed */
  blockchain: Maybe<TokenBlockchain>;
  /** Unique identifier for the blockchain */
  chainId: Scalars['String']['output'];
  /** Formatted token balance in decimals */
  formattedAmount: Maybe<Scalars['Float']['output']>;
  /** Airstack unique identifier for the data point */
  id: Scalars['ID']['output'];
  /** Block number of the latest token balance change happened */
  lastUpdatedBlock: Scalars['Int']['output'];
  /** Timestamp of the latest token balance change happened */
  lastUpdatedTimestamp: Maybe<Scalars['Time']['output']>;
  /** Nested Query allowing to retrieve address, domain names, social profiles of the owner */
  owner: Wallet;
  /** Nested Query - allows retrieving token contract level data */
  token: Maybe<Token>;
  /** Smart contract address of the token */
  tokenAddress: Scalars['Address']['output'];
  /** Unique NFT token ID */
  tokenId: Maybe<Scalars['String']['output']>;
  /** Nested Query - allows retrieving token NFT contract level data, such as images, traits, and so on */
  tokenNfts: Maybe<TokenNft>;
  /** Nested Query - allows retrieving token transfer history */
  tokenTransfers: Maybe<Array<TokenTransfer>>;
  /** Token type: ERC20, ERC721, or ERC1155 */
  tokenType: Maybe<TokenType>;
};
// ...
export enum TokenBlockchain {
  Base = 'base',
  Ethereum = 'ethereum',
  Gold = 'gold',
  Polygon = 'polygon',
  Zora = 'zora'
}
// ...
export type TokenNft = {
  /** Smart contract address of the token */
  address: Scalars['Address']['output'];
  /** Blockchain where the token smart contract is deployed */
  blockchain: Maybe<TokenBlockchain>;
  /** Unique identifier for the blockchain */
  chainId: Scalars['String']['output'];
  /** Content type of the NFT token (image, video, audio, etc.) */
  contentType: Maybe<Scalars['String']['output']>;
  /** NFT Media - resized images, animation, videos, etc. */
  contentValue: Maybe<Media>;
  /** Nested Query - allows querying the erc6551 account */
  erc6551Accounts: Maybe<Array<Account>>;
  /** Airstack unique identifier for the NFT token */
  id: Scalars['ID']['output'];
  /** Block number of the NFT token most recent transfer */
  lastTransferBlock: Maybe<Scalars['Int']['output']>;
  /** Transaction hash of the NFT token most recent transfer */
  lastTransferHash: Maybe<Scalars['String']['output']>;
  /** Timestamp of the NFT token most recent transfer */
  lastTransferTimestamp: Maybe<Scalars['Time']['output']>;
  /** NFT token metadata and attributes */
  metaData: Maybe<NftMetadata>;
  /** NFT token metadata, mirrored from the smart contract */
  rawMetaData: Maybe<Scalars['Map']['output']>;
  /** Nested Query - allows retrieving token contract level data */
  token: Maybe<Token>;
  /** Nested Query - allows querying the tokenBalance information */
  tokenBalances: Maybe<Array<TokenBalance>>;
  /** Unique NFT token ID */
  tokenId: Scalars['String']['output'];
  /** Nested Query - allows querying the tokenTransfer information */
  tokenTransfers: Maybe<Array<TokenTransfer>>;
  /** NFT token URI */
  tokenURI: Maybe<Scalars['String']['output']>;
  /** Amount of NFT tokens in the protocol */
  totalSupply: Maybe<Scalars['String']['output']>;
  /** NFT Token type: ERC721, or ERC1155 */
  type: Maybe<TokenType>;
};
// ...
export type TokenTransfer = {
  /** Token amount in the transfer */
  amount: Maybe<Scalars['String']['output']>;
  /** Token amounts in the transfer, if applicable. This mostly occurs in ERC1155 batch transfers */
  amounts: Maybe<Array<Scalars['String']['output']>>;
  /** Block number of the token transfer */
  blockNumber: Maybe<Scalars['Int']['output']>;
  /** Block timestamp of the token transfer */
  blockTimestamp: Maybe<Scalars['Time']['output']>;
  /** Blockchain where the token transfer took place */
  blockchain: Maybe<TokenBlockchain>;
  /** Unique identifier for the blockchain */
  chainId: Maybe<Scalars['String']['output']>;
  /** Formatted transfer amount in decimals */
  formattedAmount: Maybe<Scalars['Float']['output']>;
  /** Nested query - sender wallet related information, including address, domains, social profile, other token balances, and transfer history */
  from: Maybe<Wallet>;
  /** Airstack unique identifier for the data point */
  id: Maybe<Scalars['ID']['output']>;
  /** Nested query - operator wallet (if the transaction was facilitated via smart contract) related information, including address, domains, social profile, other token balances, and transfer history */
  operator: Maybe<Wallet>;
  /** Nested query - recipient wallet related information, including address, domains, social profile, other token balances, and transfer history */
  to: Maybe<Wallet>;
  /** Nested Query - allows retrieving token contract level data */
  token: Maybe<Token>;
  /** Transferred token smart contract address */
  tokenAddress: Maybe<Scalars['Address']['output']>;
  /** Unique NFT token ID */
  tokenId: Maybe<Scalars['String']['output']>;
  /** Unique NFT token IDs if multiple NFTs were a part of the transfer */
  tokenIds: Maybe<Array<Scalars['String']['output']>>;
  /** Nested Query - allows retrieving token Token NFT level data, such as images, traits, and so on for each unique NFT in the transfer */
  tokenNft: Maybe<TokenNft>;
  /** Token type: ERC20, ERC721, or ERC1155 */
  tokenType: Maybe<TokenType>;
  /** Token transfer transction hash */
  transactionHash: Scalars['String']['output'];
  /** Type of the token transfer */
  type: Maybe<Scalars['String']['output']>;
};
// ...
export enum TokenType {
  Erc20 = 'ERC20',
  Erc721 = 'ERC721',
  Erc1155 = 'ERC1155'
}
// ...
export type Wallet = {
  /** Represent On-chain smart-contract accounts */
  accounts: Maybe<Array<Account>>;
  /** Returns addresses associated with the identity input */
  addresses: Maybe<Array<Scalars['Address']['output']>>;
  /** Blockchain associated with the provided identity */
  blockchain: Maybe<TokenBlockchain>;
  /** Nested query - allows querying domains owned by the address */
  domains: Maybe<Array<Domain>>;
  /** Blockchain address, ENS domain name, social identity such as Farcaster (for Farcaster use 'fc_fid:' prefix followed by the Farcaster user ID like fc_fid:5650, or use 'fc_fname:' prefix followed by the Farcaster user ID like 'fc_fname:vbuterin') or Lens (e.g. 'stani.lens) */
  identity: Scalars['Identity']['output'];
  /** Returns Poaps owned by the address */
  poaps: Maybe<Array<Poap>>;
  /** Nested query - allows returning primary domains, if applicable */
  primaryDomain: Maybe<Domain>;
  socialFollowers: Maybe<SocialFollowerOutput>;
  socialFollowings: Maybe<SocialFollowingOutput>;
  /** Returns social profile information related to the address */
  socials: Maybe<Array<Social>>;
  /** Nested query - allows returning token balances */
  tokenBalances: Maybe<Array<TokenBalance>>;
  /** Nested query - allows returning token transfers and related information */
  tokenTransfers: Maybe<Array<TokenTransfer>>;
  /** Nested query - allows querying the XMTP enabled addresses */
  xmtp: Maybe<Array<Xmtp>>;
};
// ...
export type GetTokenBalanceQuery = { TokenBalances: { TokenBalance: Array<{ tokenAddress: any, amount: string, formattedAmount: number | null, owner: { addresses: Array<any> | null } }> | null } | null };